bhsi-backend/requirements_temp.txt
# Core web framework
fastapi>=0.104.0,<0.105.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0
uvicorn[standard]>=0.24.0,<0.25.0

# Database
sqlalchemy>=2.0.0,<2.1.0
alembic>=1.12.0,<1.13.0
# psycopg2-binary>=2.9.1  # PostgreSQL - disabled for SQLite usage

# Authentication & Security
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0
python-multipart>=0.0.5,<0.0.6
email-validator>=2.0.0,<2.1.0

# AI & NLP
google-generativeai>=0.3.0
spacy>=3.7.0,<3.9.0
# es-core-news-lg>=3.8.0  # Spanish language model for spaCy - temporarily disabled

# Advanced AI/ML for BOE - Windows-compatible versions
sentence-transformers>=2.2.0
# chromadb==0.4.24  # Temporarily disabled due to Windows compilation issues
faiss-cpu>=1.7.4  # Alternative vector database - pure Python wheels
numpy>=1.21.0
scikit-learn>=1.0.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain>=0.1.0

# HTTP & Web scraping
aiohttp>=3.8.0
httpx>=0.23.0  # Added for management summarizer
beautifulsoup4>=4.9.3
requests>=2.26.0
lxml>=4.9.0
selenium>=4.1.0
webdriver-manager>=3.8.0

# Async utilities
aiofiles>=0.8.0
tenacity>=8.0.1

# Configuration
python-dotenv>=0.19.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.18.0
httpx>=0.23.0

# Additional requirements
feedparser
aiohttp
yfinance>=0.2.0

# BigQuery dependencies
google-cloud-bigquery>=3.11.4
google-auth>=2.28.0 